Refactor honeypot logic and update header guards
Major refactor of honeypot deployment and management, including new methods for honey folder and type retrieval, and improved initialization. Removed unused cache module. Standardized header guards across all headers for consistency. Updated Evaluator logic to use new honeypot structure and improved event handling. Minor fixes and code cleanup throughout the project.

+++ RansomDetectorService/RansomDetectorService/include/file_type/pdf.cpp
@@ -14,7 +14,7 @@ namespace type_iden {
         }
 
         int corrupt_count = 0;
-        int total_objects = 0;
+        size_t total_objects = 0;
         int stream_count = 0;
         try {
             QPDF pdf;
